Call Center Dashboard
the dashboard is built to measure the performance of call center and to identify and understand which KPIs are affecting the service levels. The Dashboard answers many of client queries like why the satisfaction score is low, how many calls are answered, resolved or dropped. Apart from this, client can monitor the progress of agents monthly to be able to arrange training needed for agents. Client will also be able to see if the all or specific agents need training on specific topics to achieve the service/ satisfaction levels month-on-month.